Change crate structure for the api
