+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "osmgraph"
 authors = ["NoahSchiro"]
-version = "0.2.0"
+version = "0.3.0"
 edition = "2021"
 
 readme = "README.md"
