+++ examples/parse_graph_from_polyline.rs
@@ -1,6 +1,6 @@
 use std::error::Error;
 
-use osmgraph::overpass_api::{QueryEngine, OverpassResponse};
+use osmgraph::api::{QueryEngine, OverpassResponse};
 use osmgraph::graph::{OSMGraph, OSMNode, OSMEdge, create_graph};
 
 use serde_json::Value;
